initial test dump (still with Xcode 6, I'm on a different Mac)
